# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: GoCopyAI API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /workspaces:
    get:
      operationId: listWorkspaces
      tags: [workspaces]
      summary: List workspaces user belongs to
      responses:
        "200":
          description: Workspaces
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WorkspaceWithRole"
    post:
      operationId: createWorkspace
      tags: [workspaces]
      summary: Create a new workspace
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkspaceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceWithRole"
  /workspaces/{id}:
    get:
      operationId: getWorkspace
      tags: [workspaces]
      summary: Get workspace details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Workspace
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceDetail"
    put:
      operationId: updateWorkspace
      tags: [workspaces]
      summary: Update workspace
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkspaceInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceWithRole"
    delete:
      operationId: deleteWorkspace
      tags: [workspaces]
      summary: Delete workspace
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /workspaces/{id}/members:
    get:
      operationId: listWorkspaceMembers
      tags: [workspaces]
      summary: List workspace members
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WorkspaceMemberDetail"
    post:
      operationId: inviteWorkspaceMember
      tags: [workspaces]
      summary: Invite member by email
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InviteMemberRequest"
      responses:
        "201":
          description: Member added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceMemberDetail"
  /workspaces/{id}/members/{memberId}:
    delete:
      operationId: removeWorkspaceMember
      tags: [workspaces]
      summary: Remove member from workspace
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: memberId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed
components:
  schemas:
...
    WorkspaceWithRole:
      type: object
      required: [id, name, slug, credits, role, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        plan:
          type: string
          nullable: true
        credits:
          type: number
        logoUrl:
          type: string
          nullable: true
        primaryColor:
          type: string
          nullable: true
        role:
          type: string
          enum: [owner, admin, member]
        createdAt:
          type: string
          format: date-time
    InviteMemberRequest:
      type: object
      required: [email, role]
      properties:
        email:
          type: string
        role:
          type: string
          enum: [admin, member]
    WorkspaceInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        logoUrl:
          type: string
          nullable: true
        primaryColor:
          type: string
          nullable: true
    WorkspaceDetail:
      type: object
      required: [id, name, slug, credits, role, members, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        plan:
          type: string
          nullable: true
        credits:
          type: number
        logoUrl:
          type: string
          nullable: true
        primaryColor:
          type: string
          nullable: true
        role:
          type: string
        members:
          type: array
          items:
            $ref: "#/components/schemas/WorkspaceMemberDetail"
        createdAt:
          type: string
          format: date-time
    WorkspaceMemberDetail:
      type: object
      required: [id, userId, workspaceId, role, name, email, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        workspaceId:
          type: integer
        role:
          type: string
          enum: [owner, admin, member]
        name:
          type: string
          nullable: true
        email:
          type: string
        createdAt:
          type: string
          format: date-time